requirements.txt
torch
torchvision
tensorflow
numpy>=1.26.4,<2.0
opencv-python
Pillow
scikit-image>=0.24.0
scikit-learn
streamlit
matplotlib
